Factor out dtypes.py
